Styling fixes for mobile
