+++ packages/article-summary/src/styles/index.android.js
@@ -4,8 +4,7 @@ import sharedStyles from "./shared";
 const styles = {
   ...sharedStyles,
   labelWrapper: {
-    ...sharedStyles.labelWrapper,
-    marginBottom: spacing(1)
+    ...sharedStyles.labelWrapper
   }
 };
