+++ packages/video-label/src/style/index.js
@@ -1,23 +1,3 @@
-import { fonts } from "@times-components/styleguide";
-
-const styles = {
-  title: {
-    top: 2,
-    fontFamily: fonts.supporting,
-    fontSize: 12,
-    lineHeight: 12,
-    fontWeight: "400",
-    letterSpacing: 1.2,
-    padding: 0,
-    margin: 0,
-    position: "relative"
-  },
-  iconContainer: {
-    paddingBottom: 2
-  },
-  icon: {
-    iconSize: 8
-  }
-};
+import styles from "./shared";
 
 export default styles;
