+++ packages/video-label/src/style/index.web.js
@@ -1,0 +1,16 @@
+import sharedStyles from "./shared";
+
+const styles = {
+  title: {
+    ...sharedStyles.title
+  },
+  iconContainer: {
+    ...sharedStyles.iconContainer,
+    paddingBottom: 2
+  },
+  icon: {
+    ...sharedStyles.icon
+  }
+};
+
+export default styles;
