+++ packages/video-label/src/style/shared.js
@@ -1,0 +1,23 @@
+import { fonts } from "@times-components/styleguide";
+
+const styles = {
+  title: {
+    top: 2,
+    fontFamily: fonts.supporting,
+    fontSize: 12,
+    lineHeight: 12,
+    fontWeight: "400",
+    letterSpacing: 1.2,
+    padding: 0,
+    margin: 0,
+    position: "relative"
+  },
+  iconContainer: {
+    paddingBottom: 0
+  },
+  icon: {
+    iconSize: 8
+  }
+};
+
+export default styles;
